pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=23.3.0",
    "isort>=5.11.5",
    "langchain>=0.0.27",
    "langchain-ollama>=0.3.8",
    "python-dotenv>=0.21.1",
]

[dependency-groups]
dev = [
    "black>=23.3.0",
    "isort>=5.11.5",
]
